==> Data/0_PlayerNames.ipynb <==
<h1>Player Fetcher</h1>
<p>The purpose of this code is to scrape a list of nhl players off the hockey-reference website</p>

import pandas as pd
import numpy as np
import requests as r
import re
from string import ascii_lowercase as al #these are the lowercase ascii letters
from bs4 import BeautifulSoup as bs

<h2>Gathering the Players</h2>
<p>Before we gather any data, we need to see who we are gathering the data for.  The following code sweeps through the player inventory on <a href="https://hockey-reference.com">"hockey-reference.com</a> capturing all nhl players regardless of the year played</p>

#Main Url
url = "https://www.hockey-reference.com/players/"

nhlplayers = pd.DataFrame(columns=['player','unique_id','year_start','year_finish','position','link'])

for letter in al: #for each letter in al 
    try:
        t = r.get(url+letter) #searching for hockey players with the surname starting with letter
        soup = bs(t.text, 'html.parser')
        players = soup.find_all('p', class_='nhl') #only get the nhl players
        for i,player in enumerate(players):
            plink = player.find_all('a',href=True)
            unique_id = re.search(r'\/([\w\d\.]*)\.html',plink[0]['href'])
            play = re.search(r'([\w\.\-\s]*) \((\d{4})-(\d{4}), ([\w\b]*)',player.text)
            inputdata = [play[1],unique_id[1],play[2],play[3],play[4],'https://www.hockey-reference.com'+plink[0]['href']]
            #print(i,inputdata)
            nhlplayers = nhlplayers.append(dict(zip(nhlplayers.columns, inputdata)), ignore_index=True)
        print("Finished adding %d players for letter %s" % (i,letter))
    except:
        print(plink)

<h2>Filtering the Players</h2>
<p>While we previously filtered only players in the NHL, we now need to refine the filtering slightly, removing players who played before the availablility of game logs).  Initially, I'll only keep players who played after 1990 in the hopes that I can get the injury data for them.  This filtering drops the number of available players from 7983 to 4842.  The other thing I'll do for simplification is to change the reduce number of positions to Forward, Defense or Goalie.</p>

#convert datatypes as appropriate
nhlplayers[['year_start','year_finish']] = nhlplayers[['year_start','year_finish']].astype(int)

nhlplayers.info()

#drop players who finished before 1990
nhlplayers.drop(nhlplayers[nhlplayers['year_finish']<1990].index,inplace=True)

nhlplayers.info()

nhlplayers['position'].unique()

position_remap = {'C':'F','LW':'F','RW':'F','D':'D','G':'G','W':'F'}
nhlplayers['position'] = nhlplayers['position'].replace(position_remap)
nhlplayers['position'].unique()

nhlplayers[nhlplayers['position']=='G'].reset_index().to_csv('nhlplayerlist_goalieonly.txt','\t')

nhlplayers[nhlplayers['position']=='G']
