==> Data/1_GetPlayerGamelogs.ipynb <==
<h1>Gamelog Grabber</h1>
<p>The following code scrapes the game logs for the 4000+ players in the inital directory and saves them to individual csvs.  Only seasons from the last 30 years (from 1990-1991 onwards) are considered.</p>

import pandas as pd
import numpy as np
import requests as r
import re
from bs4 import BeautifulSoup as bs

pd.set_option('display.max_rows', 50)
pd.set_option('display.max_columns', 50)
pd.set_option('display.width', 1000)

#import player list from csv and tidy 
players = pd.read_csv('nhlplayerlist_goalieonly.txt','\t')
players.drop(columns=['Unnamed: 0','index'],axis=1,inplace=True)
players.head(5)

<h3>Functions to aid in Data Scrapping</h3>
<p>The what follows are the functions used in the data collection<ul><li><p><strong>loglink</strong> - Gets the urls for each year the player played and returns it as a dictionary</p></li><li><p><strong>gamelogger</strong> - Grabs the game logs for each of the links entered and returns them as a DataFrame.</p></li></ul>
</p>


def loglink(pwebsite):
    """
    Gets the urls for each year the player played and returns it as a dictionary
    input: (string) url for the player in question
    output: (dict) year/link pairs for each year the player competed
    """
    page = r.get(pwebsite)
    soup = bs(page.text,'html.parser')
    links = soup.find_all('a',href=True)

    #find the game log links
    gamelogs = [] #the list of the game log inks
    for l in links:
        if 'gamelog' in l['href']:
            gamelogs.append(l['href'])
    gamelogs = list(dict.fromkeys(gamelogs)) #remove duplicate entries

    #find the game logs
    gameloglinks = {} #a dictionary with the year:gameloglink
    for gl in gamelogs:
        reg = re.search(r'gamelog\/([\d]*)',gl)
        if reg[1]:
            gameloglinks[int(reg[1])] = gl
    return gameloglinks
def gamelogger(loglinks):
    """
    Grabs the game logs for each of the links entered and returns them as a DataFrame
    Input:  (dict) Year/Gamelog pairs
    Output: (dataframe) the entire post-1990 gamelog data (excluding playoffs) for that player
    """
    glogs = pd.DataFrame()
    #for every year after 1990 (1991 is the 1990-1991 season), grab the gamelogs
    column_names = []
    for key,value in loglinks.items():
        if key>1990:
            page = r.get('https://hockey-reference.com'+value)
            soup = bs(page.text,'html.parser')

            for eachitem in soup.find_all('tr',id=True): #for each row in table
                rowdata = []
                for eachc in eachitem.find_all('td'): #for each column in row
                    if glogs.empty:
                        column_names.append(eachc['data-stat'])
                    rowdata.append(eachc.text)
                if glogs.empty:
                    glogs = pd.DataFrame(columns=column_names)
                glogs = glogs.append(dict(zip(glogs.columns,rowdata)), ignore_index=True) 
    return glogs

<h3>Get the data</h3>
<p>Data is aggregated and saved in tabs delimited format where each player is a different file and the name is the uniqueid.</p>

failed_players = [] #capture failed players
for i,player in players.iterrows():
    try:
        gamelogdict = loglink(player['link'])
    except:
        failed_players.append(player['player'])
        print("Failed to get gamelog links for %s" % (player['player']))
    try:
        log = gamelogger(gamelogdict)
    except:
        failed_players.append(player['player'])
        print("Error getting the gamelog for %s" % (player['player']))
    #write gamelog to csv for further analysis    
    log.to_csv('player_gamelogs\\'+player['unique_id']+'.txt')
    print('finished %d of %d or %.2f%%'%(i,players.shape[0],100*i/players.shape[0]))

#fix dead link for 3 failed players 
#failed_players=['Todd Reirden','Erik Reitz','Mike Vernon']
for p in failed_players:
    gamelogd = loglink(players.loc[players['player']==p,'link'].values[0])
    log = gamelogger(gamelogd)
    uid = players.loc[players['player']==p,'unique_id'].values[0]
    print(uid)
    log.to_csv('player_gamelogs\\'+uid+'.txt')
