==> Data/4_PlayerData.ipynb <==
<h1>Get Player Data</h1>
<p>Get player weights and heights and total number of games in the nhl</p>

import pandas as pd
import numpy as np
import requests as r
import re
from bs4 import BeautifulSoup as bs

pd.set_option('display.max_rows', 50)
pd.set_option('display.max_columns', 50)
pd.set_option('display.width', 1000)

#Main Url
#import player list from csv and tidy 
players = pd.read_csv('nhlplayerlist_goalies.txt','\t')
players.drop(columns=['Unnamed: 0'],axis=1,inplace=True)
players.head(5)

<p>Add columns for height and weight</p>

players['height'] = np.nan
players['weight'] = np.nan

for i,player_link in enumerate(players['link']):
    t = r.get(player_link) #searching for hockey players with the surname starting with letter
    soup = bs(t.text, 'html.parser').text
    
    #extract the info from the page
    try:
        players.loc[i,'height_cm'] = re.search(r'(\d{3})cm',soup)[1]
        players.loc[i,'weight_kg'] = re.search(r'(\d{2,3})kg',soup)[1]
    except: #leave as nan if not found
        print(players.loc[i,'player'],players.loc[i,'height_cm'],players.loc[i,'weight_kg'])

players['total_games'] = np.nan
for row,p in enumerate(players['unique_id']):
    t = pd.read_csv('player_gamelogs\\'+p+'.txt','\t')
    players.loc[row,'total_games']= t.shape[0]

players.to_csv('nhlplayerlist_goalieswhwtg.txt','\t')
