==> Data/5_TeamLocation.ipynb <==
<h2>Gets Team Location and Timezone</h2>
<p>put together a list of the teams with the following information<ul><li>team_abbr</li><li>team_name</li><li>years_active</li><li>city</li><li>latitude</li><li>longitude</li><li>timezone</li></ul></p>

import pandas as pd
import numpy as np
import requests as r
import re
from bs4 import BeautifulSoup as bs

pd.set_option('display.max_rows', 50)
pd.set_option('display.max_columns', 50)
pd.set_option('display.width', 1000)

#load gamelog list

teamDataFrame = pd.read_csv('team_gamelog_list.txt','\t')
teamDataFrame.drop(columns=['Unnamed: 0'],inplace=True)
teamDataFrame

#list the unique abbreviations:
abbrs = teamDataFrame['team_abbr'].unique()

teams = pd.DataFrame(columns=['team_abbr','team_name','years_active','city','latitude','longitude','timezone'])
for r,abr in enumerate(abbrs):
    #input abr
    teams.loc[r,'team_abbr'] = abr
    #input team name
    teams.loc[r,'team_name'] = teamDataFrame[teamDataFrame['team_abbr']==abr]['team_name'].unique()[0]
    #input years active
    ya = teamDataFrame[teamDataFrame['team_abbr']==abr]['Season'].str[:4].astype(int).describe()
    teams.loc[r,'years_active'] = str(int(ya['min']))+'-'+str(int(ya['max'])+1)

#fill in the rest by hand
teams.to_csv('team_list.csv',',')
